update bank client structs to adhere to the model convention
